pinterest-scraper/requirements.txt
scrapy-playwright==0.0.33
beautifulsoup4==4.12.2
lxml==4.9.3